+++ android/src/main/java/io/bluedot/BluedotPointSdkModule.java
@@ -310,6 +310,16 @@ public class BluedotPointSdkModule extends ReactContextBaseJavaModule implements
                         location.putDouble("longitude", loc.getLongitude());
 
                         destination.putMap("location", location);
+                        
+                        if (destinationObj.getCustomData() != null) {
+                            WritableMap customData = new WritableNativeMap();
+                            Map<String, String> customDataMap = destinationObj.getCustomData();
+                            for (Map.Entry<String, String> entry : customDataMap.entrySet()) {
+                                customData.putString(entry.getKey(), entry.getValue());
+                            }
+                            destination.putMap("customData", customData);
+                        }
+                        
                         zone.putMap("destination", destination);
                     }
                     zoneList.pushMap(zone);